requirements.txt
fastapi
uvicorn
python-dotenv
requests
pydub
twilio
websockets
aiofiles
numpy
soundfile
ffmpeg-python